requirements.txt
# MCP SDK
mcp>=0.1.0

# YAML processing
pyyaml>=6.0

# HTTP requests (for some tools)
requests>=2.31.0

# Encoding/decoding
chardet>=5.0.0

# Additional dependencies for Kali tools integration
mako>=1.3.0
prettytable>=3.9.0
jinja2>=3.1.0
paramiko>=3.3.0
cryptography>=41.0.0
pillow>=10.0.0
